// apps/everest/pages/new-database/steps/third/third-step.types.ts
export enum TimeValue {
  hours = 'hours',
  days = 'days',
  weeks = 'weeks',
  months = 'months',
}

export enum StorageLocation {
  S3 = 'S3',
  Local = 'Local',
}

export enum WeekDays {
  Mo = 'Monday',
  Tu = 'Tuesday',
  We = 'Wednesday',
  Th = 'Thursday',
  Fr = 'Friday',
  Sa = 'Saturday',
  Su = 'Sunday',
}

export enum AmPM {
  AM = 'AM',
  PM = 'PM',
}